backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
requests
motor==3.4.0
pymongo==4.6.0
python-jose[cryptography]
passlib>=1.7.4
bcrypt==4.0.1
python-multipart
groq
numpy
PyJWT
email-validator
